Revert "Keep applying 1 and 2 SOLID principles"
This reverts commit 2f140a37dbe0f9f56e5b91df2887a4a4558c0e93.
