--- src/Components/payment/dataInfo.js
@@ -1,16 +1,0 @@
-export const dataPayment = {
-  cardNumber: cardNumber,
-  expYear: `20${expDate.split("/")[1]}`,
-  expMonth: `${expDate.split("/")[0]}`,
-  CVC: CVC,
-  docType: docType,
-  docNumber: docNumber,
-  bill: "OR-1234",
-  description: "Quick Suscription Payment",
-  value: "9",
-  tax: "1",
-  taxBase: "8",
-  currency: "USD",
-  dues: "1",
-  ip: "190.000.000.000",
-};
